third_party/blink/web_tests/external/wpt/css/css-fonts/variations/font-style-parsing-expected.txt

This is a testharness.js-based test.
[FAIL] Font-style (computed): 'oblique' is valid
  assert_equals: Font-style computed style: 'oblique' is valid expected "oblique" but got "italic"
[FAIL] Font-style (computed): 'oblique' followed by zero degrees is valid
  assert_equals: Font-style computed style: 'oblique' followed by zero degrees is valid expected "oblique 0deg" but got "normal"
[FAIL] Font-style (computed): 'oblique' followed by positive angle in radians is valid
  assert_regexp_match: Font-style computed style: 'oblique' followed by positive angle in radians is valid expected object "/^oblique 28\\.64\\d*deg$/" but got "oblique 28.5deg"
[FAIL] Font-style (computed): 'oblique' followed by calc is valid even if it mixes units (with computation)
  assert_equals: Font-style computed style: 'oblique' followed by calc is valid even if it mixes units (with computation) expected "oblique" but got "italic"
Harness: the test ran to completion.
